+++ packages/stack/src/views/GestureHandler.windows.tsx
@@ -2,5 +2,5 @@ export {
   PanGestureHandler,
   GestureHandlerRootView,
   GestureState,
-} from './GestureHandlerCompat';
+} from './GestureHandlerStub';
 export type { PanGestureHandlerGestureEvent } from 'react-native-gesture-handler';
