+++ packages/stack/src/views/MaskedView.tsx
@@ -1,10 +1,1 @@
-import type * as React from 'react';
-
-type Props = {
-  maskElement: React.ReactElement;
-  children: React.ReactElement;
-};
-
-export default function MaskedView({ children }: Props) {
-  return children;
-}
+export { default } from './MaskedViewStub';
