+++ packages/stack/src/views/MaskedView.windows.tsx
@@ -1,3 +1,1 @@
-import StubbedMaskedView from './MaskedView';
-
-export default StubbedMaskedView;
+export { default } from './MaskedViewStub';
