+++ packages/stack/src/views/MaskedViewStub.tsx
@@ -1,0 +1,10 @@
+import type * as React from 'react';
+
+type Props = {
+  maskElement: React.ReactElement;
+  children: React.ReactElement;
+};
+
+export default function MaskedView({ children }: Props) {
+  return children;
+}
